LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/the-offscreen-canvas/2d.canvas.readonly-expected.txt

2d.canvas.readonly

canvas is readonly


PASS canvas is readonly 
